+++ packages/web/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @powersync/web
 
+## 1.38.6
+
+### Patch Changes
+
+- Updated dependencies [115d736]
+- Updated dependencies [bd0bc41]
+  - @powersync/common@1.57.0
+
+## 1.38.5
+
+### Patch Changes
+
+- Updated dependencies [1c6e980]
+- Updated dependencies [17e23df]
+  - @powersync/common@1.56.0
+
 ## 1.38.4
 
 ### Patch Changes