+++ tools/diagnostics-app/CHANGELOG.md
@@ -1,5 +1,19 @@
 # diagnostics-app
 
+## 0.13.13
+
+### Patch Changes
+
+- @powersync/react@1.10.0
+- @powersync/web@1.38.6
+
+## 0.13.12
+
+### Patch Changes
+
+- @powersync/react@1.10.0
+- @powersync/web@1.38.5
+
 ## 0.13.11
 
 ### Patch Changes